Add the ability for a Drop to have an action.
This adds the ability for a Drop to execute a closure when tapped. In addition
to the action completion block, it also has a `userInfo` dictionary which can
allow a user to pass along data to work with when the closure executes, such as
when using a Drop to respond to a Push Notification.